Port to Dune 3
This enables us to use the formatter and also moves the dependencies to
the `dune-project` file which will generate the OPAM file.

It allows us to use `implicit_transitive_deps` using `-H` on supported
compilers in the future which should avoid the issues with the previous
implementation without compiler support.