Use JaneStreet `sexp` instead of `sexpdata`
The issue is that `sexpdata` parses numbers, so something like `(dune
lang 1.10)` is parsed as floating point and reduced to `(dune lang 1.1)`
which turns the dune-file invalid in some cases.

`sexp` is an OCaml tool from Jane Street, unfortunately it doesn't allow
matching and deleting, only matching and rewriting so we rewrite using
`sexp` and delete using `grep` in an additional step.