Use Dune to determine depexts
The issue here is that OPAM does not install depopts by default and even
despite #102 picking depopts explicitely this only works if the depopts
are in the repo. E.g. if a package depends on a package that has a
depopt itself, it won't be detected by #102 and OPAM won't suggest the
`depext`. However, Dune will pick it when locking and then fail.

Thus this code changes the logic, so Dune locks a solution first and is
then used to output the packages that need to be installed.